fixed a nasty buffering bug --Ths lisne, and those below, will be ignored--
M    trunk/csharp/core/media/data/audio/PCMFormatInfo.cs
M    trunk/csharp/core/media/data/audio/PCMDataInfo.cs
M    trunk/csharp/core/media/data/audio/codec/WavAudioMediaData.cs
M    trunk/csharp/core/media/data/DataProviderManager.cs
M    trunk/csharp/core/media/data/FileDataProvider.cs
M    trunk/csharp/core/publish/AbstractPublishManagedAudioVisitor.cs
M    trunk/csharp/XukImport/DaisyToXuk/DaisyToXuk2.cs

+++ csharp/core/media/data/FileDataProvider.cs
@@ -292,7 +292,7 @@ namespace urakawa.media.data
                 {
                     if (thisData.Length > 0)
                     {
-                        data.DataProviderManager.AppendDataToProvider(thisData, (int) thisData.Length, c);
+                        data.DataProviderManager.AppendDataToProvider(thisData, thisData.Length, c);
                     }
                 }
                 finally
@@ -318,7 +318,7 @@ namespace urakawa.media.data
                 {
                     if (thisStm.Length > 0)
                     {
-                        DataProviderManager.AppendDataToProvider(thisStm, (int) thisStm.Length, expFDP);
+                        DataProviderManager.AppendDataToProvider(thisStm, thisStm.Length, expFDP);
                     }
                 }
                 finally
